+++ src/main.ts
@@ -48,6 +48,7 @@ import * as preNPCUpdate from "./callbacks/preNPCUpdate";
 import * as preRoomEntitySpawn from "./callbacks/preRoomEntitySpawn";
 import * as preSpawnClearAward from "./callbacks/preSpawnClearAward";
 import * as preUseItem from "./callbacks/preUseItem";
+import * as postAmbushStarted from "./callbacksCustom/postAmbushStarted";
 import * as postBombExploded from "./callbacksCustom/postBombExploded";
 import * as postBoneSwing from "./callbacksCustom/postBoneSwing";
 import * as postCursedTeleport from "./callbacksCustom/postCursedTeleport";
@@ -161,6 +162,7 @@ function registerCallbacksVanilla(mod: ModUpgraded) {
 }
 
 function registerCallbacksCustom(mod: ModUpgraded) {
+  postAmbushStarted.init(mod);
   postBombExploded.init(mod);
   postBoneSwing.init(mod);
   postCursedTeleport.init(mod);